+++ skeleton/app/user/main.php
@@ -14,16 +14,16 @@ class UserController extends C_Controller
 
     public function _profile()
     {
-        $user_id = $input->getInt('user_id');
-        if ( $user_id == false ) {
-            return json_view(STATUS_INVALID_ARGS, 'Invalid Arguments');
+        $sess = $this->isLoggedIn();
+        if ( $sess == false ) {
+            return json_view(STATUS_NO_SESSION, 'No session');
         }
 
         $data = array(
-            'user_id'   => $user_id,
-            'head_img'  => 'http://git.oschina.net/uploads/87/5187_lionsoul.jpg',
-            'nickname'  => 'lionsoul',
-            'signature' => '平凡 | 执着'
+            'user_id'   => $sess->getUid(),
+            'head_img'  => $sess->get('head_img'),
+            'nickname'  => $sess->get('nickname'),
+            'signature' => $sess->get('signature')
         );
 
         return json_view(STATUS_OK, $data);
@@ -31,13 +31,32 @@ class UserController extends C_Controller
     
     public function _signIn($input)
     {
-        $data = array(
-            'head_img'  => 'http://git.oschina.net/uploads/87/5187_lionsoul.jpg',
-            'nickname'  => 'lionsoul',
-            'signature' => '平凡 | 执着'
-        );
+        import('Session', false);
 
-        return json_view(STATUS_OK, $data);
+        $sessKey = isset($this->session_key) ? $this->session_key : 'File';
+        $conf    = config("session#{$sessKey}");
+        $conf['r8c']    = true;
+        $conf['sessid'] = '1707ydlteVx9VPM9W0LBMHKCRWT648m9';
+
+        $sessObj = Session::start($sessKey, $conf);
+        $sessObj->register(7);
+
+        $sessObj->set('user_id',  7);
+        $sessObj->set('head_img', 'http://git.oschina.net/uploads/87/5187_lionsoul.jpg');
+        $sessObj->set('nickname', 'lionsoul');
+        $sessObj->set('signature', '平凡 | 执着');
+
+        return json_view(STATUS_OK, 'Ok');
+    }
+
+    public function _logOut()
+    {
+        $sess = $this->isLoggedIn();
+        if ( $sess != false ) {
+            $sess->destroy();
+        }
+
+        return json_view(STATUS_OK, 'Ok');
     }
 
     public function _signUp($input)